refactor(packages/explorer/src/views/Mining/tokenMiner.js packages/explorer/src/views/Mining/minePro
